refactor: custom page to hooks
