+++ vite-project/src/components/tiles-list/tiles-list.tsx
@@ -3,6 +3,7 @@ import Tile from '../tile/tile';
 import NodeType from '../../types/node';
 
 export default function TilesList(props: { nodes: NodeType[] }) {
+  console.log(1);
   return (
     <div className="tiles">
       {props.nodes.map((node: NodeType) => (
