+++ vite-project/src/pages/custom/Custom.tsx
@@ -1,30 +1,19 @@
-import React from 'react';
-import AddForm from '../../components/add-card-form/Form';
-import NewForm from '../../components/add-card-form/new-form';
+import React, { useState } from 'react';
+import AddForm from '../../components/add-card-form/Form.tsx';
 import TilesList from '../../components/tiles-list/tiles-list';
 import NodeType from '../../types/node';
 
-export default class Custom extends React.Component<object, { nodes: NodeType[] }> {
-  resData: NodeType[] = [];
+export default function Custom() {
+  const [nodesList, useNodes] = useState<NodeType[]>([]);
 
-  constructor(props: object | Readonly<object>) {
-    super(props);
-    this.state = { nodes: [] };
-    this.handleFormSubmit = this.handleFormSubmit.bind(this);
+  function HandleFormSubmit(data: NodeType) {
+    useNodes((nodesList) => [...nodesList, data]);
   }
 
-  handleFormSubmit(data: NodeType) {
-    this.resData.push(data);
-    this.setState({ nodes: this.resData });
-  }
-
-  render() {
-    return (
-      <>
-        <NewForm onFormSubmit={this.handleFormSubmit} />
-        {/* <AddForm onFormSubmit={this.handleFormSubmit} /> */}
-        <TilesList nodes={this.state.nodes} />
-      </>
-    );
-  }
+  return (
+    <>
+      <AddForm onFormSubmit={HandleFormSubmit} />
+      <TilesList nodes={nodesList} />
+    </>
+  );
 }
